refactor(avatar): remove dead procedural engine, migrate remaining callers
knip flags the retained shape/eye/nose engine as dead once boring-avatars
took over. Delete avatar-parts.tsx, avatar-generator.tsx, and
gradient-from-seed.ts (+ their tests); keep the two survivors
(isPhotoAvatarUrl, AvatarDraft) in a small photo.ts module.

Migrate the last procedural callers (studios/recruit routes,
scenario-presets) to randomBeamAvatar().
