+++ src/web/src/components/avatar/agent-avatar.test.ts
@@ -1,8 +1,11 @@
 import { describe, it, expect } from "vitest"
 import { AgentAvatar } from "./agent-avatar"
 import { BoringAvatar } from "./boring-avatar"
-import { serializeAvatarConfig, DEFAULT_CONFIG } from "./avatar-parts"
 import { serializeBeamSeed } from "@/lib/avatar/seed-url"
+
+// A legacy procedural config value (the format the removed engine used to
+// store) — the renderer must ignore it and fall back to an id-seeded beam.
+const LEGACY_CONFIG = 'avatar:{"shape":"book","eye":"happy","nose":"dash","bg":1}'
 
 type ImgEl = { type: "img"; props: { src: string; alt: string; style: { width: number; height: number } } }
 type BeamEl = { type: typeof BoringAvatar; props: { seed: string; size: number } }
@@ -33,8 +36,7 @@ describe("AgentAvatar", () => {
   })
 
   it("ignores a legacy avatar:{shape…} config and beams by the fallback seed", () => {
-    const url = serializeAvatarConfig(DEFAULT_CONFIG)
-    const el = AgentAvatar({ name: "Bot", avatarUrl: url, seed: "agent-1", size: 32 }) as unknown as BeamEl
+    const el = AgentAvatar({ name: "Bot", avatarUrl: LEGACY_CONFIG, seed: "agent-1", size: 32 }) as unknown as BeamEl
     expect(el.type).toBe(BoringAvatar)
     expect(el.props.seed).toBe("agent-1")
   })
