+++ src/web/src/components/avatar/index.ts
@@ -1,14 +1,4 @@
-export {
-  type AvatarConfig,
-  type AvatarDraft,
-  AvatarRenderer,
-  DEFAULT_CONFIG,
-  serializeAvatarConfig,
-  parseAvatarUrl,
-  randomConfig,
-  configFromName,
-  isPhotoAvatarUrl,
-} from "./avatar-parts";
+export { type AvatarDraft, isPhotoAvatarUrl } from "./photo";
 
 export { BoringAvatar } from "./boring-avatar";
 export { MarbleBackground } from "./marble-background";
